completed the 2nd assignment, implementing different GET/POST method

+++ src/utils.cc
@@ -3,7 +3,8 @@
 #include <cstring>
 #include <fcntl.h>
 #include <iostream>
-#include <syslog.h> 
+#include <syslog.h>
+#include <unistd.h> 
 
 #include "utils.hh"
 
@@ -26,7 +27,7 @@ void usage(const char* prog)
 // - std::stoi()
 int16_t select_port(int argc, char **argv)
 {
-  int16_t server_port_tmp = DEFAULT_PORT;
+  int server_port_tmp = DEFAULT_PORT;
   if (argc > 2) {		// Too many parameters
     usage(argv[0]);		// No return
   }
